+++ source/Energinet.DataHub.Charges.Libraries/source/Energinet.DataHub.Charges.Clients/documents/release-notes/release-notes.md
@@ -2,7 +2,8 @@
 
 ## Version 4.3.2
 
-Added 'Utc' to dates in `ChargePricesSearchCriteriaV1Dto`
+Added 'Utc' to dates in `ChargePricesSearchCriteriaV1Dto`.
+Renamed `SortColumnName` to `ChargePriceSortColumnName`.
 
 ## Version 4.3.1